requirements.txt
asgiref==3.5.2
Django==4.1.3
django-ckeditor==6.5.1
django-js-asset==2.0.0
psycopg2==2.9.6
sqlparse==0.4.3
